test: unskip metrics tests
